Fix import and move files into folders
