+++ README.md
@@ -1,3 +1,9 @@
 # mdn-storybook
 
 UI components for MDN Web Docs
+
+## Getting Started
+
+1. Run `npm install` to install dependencies
+2. Run `npm run storybook` to start Storybook
+3. Visit http://localhost:9009 in your browser
