+++ src/components/Logo/Logo.stories.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 
-import Logo from "./logo.jsx";
+import Logo from "./Logo.jsx";
 
 storiesOf("Logo", module)
   .add("default", () => <Logo locale="en-US" />)
