+++ src/components/MainNav/MainNav.stories.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { storiesOf } from "@storybook/react";
 
-import MainNav from "./main-nav.jsx";
+import MainNav from "./MainNav.jsx";
 
 storiesOf("MainNav", module)
   .add("default", () => <MainNav locale="en-US" />)
